# PZ/calculator/server/swagger_server/models/result.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Result(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, result: int=None):  # noqa: E501
        """Result - a model defined in Swagger

        :param result: The result of this Result.  # noqa: E501
        :type result: int
        """
        self.swagger_types = {
            'result': int
        }

        self.attribute_map = {
            'result': 'result'
        }
        self._result = result

    @classmethod
    def from_dict(cls, dikt) -> 'Result':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The result of this Result.  # noqa: E501
        :rtype: Result
        """
        return util.deserialize_model(dikt, cls)

    @property
    def result(self) -> int:
        """Gets the result of this Result.


        :return: The result of this Result.
        :rtype: int
        """
        return self._result

    @result.setter
    def result(self, result: int):
        """Sets the result of this Result.


        :param result: The result of this Result.
        :type result: int
        """
        if result is None:
            raise ValueError("Invalid value for `result`, must not be `None`")  # noqa: E501

        self._result = result